+++ src/script_curator/newsletter_script.ts
@@ -1,8 +1,10 @@
-import { curateAndGenerateNewsletter } from './newletter_txt_format';
+import { curateAndGenerateNewsletter } from './newsletter_txt_format';
 
 async function runNewsletter() {
     const newsletterText = await curateAndGenerateNewsletter();
     console.log(newsletterText);  // Affiche ou traite la newsletter ici
 }
+
+
 
 runNewsletter();
