+++ src/script_curator/newsletter_txt_format.ts
